+++ ReSwift/CoreTypes/StoreSubscriber.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2015 ReSwift Community. All rights reserved.
 //
 
-public protocol AnyStoreSubscriber: class {
+public protocol AnyStoreSubscriber: AnyObject {
     // swiftlint:disable:next identifier_name
     func _newState(state: Any)
 }