+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+**Other:**
+- Make isDispatching of Store atomic (#341) - @zhongwuzw
+
 # 5.0.0
 
 *Released: 2019-06-30*