+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Upcoming
 
+**Breaking API Changes:**
+- Drop support for Swift 3.2, 4.0, and 4.1. (#418) - @DivineDominion
+
 **API Changes:**
-- Add capability to mutate `Middleware` after store creation. (#) - @mjarvis
+- Add capability to mutate `Middleware` after store creation. (#427) - @mjarvis
 
 **Other:**
 - Make isDispatching of Store atomic (#341) - @zhongwuzw