+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+<!--
 # Upcoming
+
+**Breaking API Changes:**
+**API Changes:**
+**Other:**
+-->
+
+# 6.0.0
+
+*Released: 2020-10-01*
 
 **Breaking API Changes:**
 - Drop support for Swift 3.2, 4.0, and 4.1. (#418) - @DivineDominion